Collapse idp/ sub-package into single discover/idp.go
Follows the existing pattern (asn.go, cdn.go) of keeping simpler
commands as single files in the discover package rather than
sub-directories.
